backend/requirements_verba.txt
goldenverba>=0.2.0
python-dotenv
weaviate-client>=4.4.0
openai>=1.6.0
flask-cors>=4.0.0
